refac: 검색 결과 방향키 custom hook으로 분리
